C++ coding Sol.  4.  Passage to the Other World
Based on the text file routine I made a simple coding program that can be used to see the schedule of students class. For that first you need to enter the 1 or 2 or 3 . 
1 is for first year , 2 is for second year and 3 is for third year. If you enter another any number it will show you default case you do not have any class. 
A simple algorithm was prepared manually. there are 3 room necessary to operate all the classes. But little sifting is nesessary while pc pool/ lab has a different setup.
we can also add more functionality and algorithms if we want to make it more flexible.